[racingcar]refactor: rename variable and remove duplicated method call
