+++ src/main/java/racingcar/domain/Car/CarNames.java
@@ -4,22 +4,26 @@ import java.util.Arrays;
 
 public class CarNames {
     private static final String SPLIT_REGEX = ",";
-    private final String carNames;
+    private final String[] parsed;
 
     public CarNames(final String value) {
-        validate(value);
-        this.carNames = value;
+        this.parsed = parse(value);
+        validate();
     }
 
-    private void validate(final String value) {
-        if (value.split(SPLIT_REGEX).length == 0) {
+    private String[] parse(final String value) {
+        return Arrays.stream(value.split(SPLIT_REGEX))
+                .map(s -> s.trim())
+                .toArray(String[]::new);
+    }
+
+    private void validate() {
+        if (parsed.length == 0) {
             throw new RuntimeException("자동차 이름이 존재하지 않습니다");
         }
     }
 
-    public String[] parse() {
-        return Arrays.stream(carNames.split(SPLIT_REGEX))
-                .map(s -> s.trim())
-                .toArray(String[]::new);
+    public String[] getParsed() {
+        return Arrays.copyOf(parsed, parsed.length);
     }
 }
